backend/requirements.txt
# ============================================
# Market Path Finder App - Backend Dependencies
# Target: Raspberry Pi 5 + 4070 Ti PC
# ============================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
# psycopg2-binary: Linux (Raspberry Pi) only

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# AI / LLM
langchain==0.3.14
langchain-community==0.3.14
mcp==1.3.0

# MQTT
paho-mqtt==2.1.0

# QR Code
qrcode[pil]==8.0

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Utilities
httpx==0.28.1
python-dotenv==1.0.1

# Path Algorithm
networkx==3.4.2
# numpy: add in Step 3 (requires Python 3.11/3.12 or Raspberry Pi)

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
